backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
pyproj==3.6.1
geopy==2.4.1
torch==2.3.1
scikit-learn==1.5.0
python-dotenv==1.0.1
pydantic==2.7.4